+++ src/utils/merge-queue.ts
@@ -12,9 +12,18 @@ limitations under the License.
 */
 
 import { paginateAllBranches } from './paginate-all-branches';
+import { context } from '@actions/github';
 
 export const getMergeQueueCommitHashes = async () => {
   const branches = await paginateAllBranches();
   const mergeQueueBranches = branches.filter(branch => branch.name.startsWith('gh-readonly-queue/'));
   return mergeQueueBranches.map(branch => branch.commit.sha);
 };
+
+export const getPrNumberFromMergeQueueRef = () =>
+  Number(
+    context.ref
+      .split('/')
+      .find(part => part.includes('pr-'))
+      ?.match(/\d+/)?.[0]
+  );
